Optimizations and cleanup
