+++ 3D_Game_Engine/3D_Game_Engine/3D_Game_Engine.vcxproj
@@ -88,16 +88,15 @@
     <ClCompile Include="Core\InputManager.cpp" />
     <ClCompile Include="Core\IOManager.cpp" />
     <ClCompile Include="Core\ViewPort.cpp" />
+    <ClCompile Include="JML\Math\JadeMath.cpp" />
+    <ClCompile Include="JML\Math\mat3.cpp" />
+    <ClCompile Include="JML\Math\mat4.cpp" />
+    <ClCompile Include="JML\Math\quat.cpp" />
+    <ClCompile Include="JML\Math\vec2.cpp" />
+    <ClCompile Include="JML\Math\vec3.cpp" />
+    <ClCompile Include="JML\Math\vec4.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="MainGame.cpp" />
-    <ClCompile Include="Math\JadeCalculator.cpp" />
-    <ClCompile Include="Math\JadeMath.cpp" />
-    <ClCompile Include="Math\mat3.cpp" />
-    <ClCompile Include="Math\mat4.cpp" />
-    <ClCompile Include="Math\quat.cpp" />
-    <ClCompile Include="Math\vec2.cpp" />
-    <ClCompile Include="Math\vec3.cpp" />
-    <ClCompile Include="Math\vec4.cpp" />
     <ClCompile Include="Rendering\AssImpModelLoader.cpp" />
     <ClCompile Include="Rendering\Lighting.cpp" />
     <ClCompile Include="Rendering\Mesh.cpp" />
@@ -138,17 +137,16 @@
     <ClInclude Include="Core\InputManager.h" />
     <ClInclude Include="Core\IOManager.h" />
     <ClInclude Include="Core\ViewPort.h" />
+    <ClInclude Include="JML\Math\JadeMath.h" />
+    <ClInclude Include="JML\Math\Jade_Primitives.h" />
+    <ClInclude Include="JML\Math\JML.h" />
+    <ClInclude Include="JML\Math\mat3.h" />
+    <ClInclude Include="JML\Math\mat4.h" />
+    <ClInclude Include="JML\Math\quat.h" />
+    <ClInclude Include="JML\Math\vec2.h" />
+    <ClInclude Include="JML\Math\vec3.h" />
+    <ClInclude Include="JML\Math\vec4.h" />
     <ClInclude Include="MainGame.h" />
-    <ClInclude Include="Math\JadeCalculator.h" />
-    <ClInclude Include="Math\JadeMath.h" />
-    <ClInclude Include="Math\Jade_Primitives.h" />
-    <ClInclude Include="Math\JML.h" />
-    <ClInclude Include="Math\mat3.h" />
-    <ClInclude Include="Math\mat4.h" />
-    <ClInclude Include="Math\quat.h" />
-    <ClInclude Include="Math\vec2.h" />
-    <ClInclude Include="Math\vec3.h" />
-    <ClInclude Include="Math\vec4.h" />
     <ClInclude Include="Rendering\AssImpModelLoader.h" />
     <ClInclude Include="Rendering\Lighting.h" />
     <ClInclude Include="Rendering\Mesh.h" />
