+++ 3D_Game_Engine/3D_Game_Engine/JML/Math/JML.h
@@ -1,0 +1,21 @@
+// ***********************************************************************
+// Author           : Jesse Derochie
+// Created          : 03-23-2016
+// Created For      : Jade Math Library
+// Last Modified By : Jesse Derochie
+// Last Modified On : 03-24-2016
+// ***********************************************************************
+// <copyright file="JML.h">
+//     Copyright (c) Jesse Derochie. All rights reserved.
+// </copyright>
+// <summary>JML Math Library header file</summary>
+// ***********************************************************************
+
+#include "Jade_Primitives.h"
+#include "JadeMath.h"
+#include "vec2.h"
+#include "vec3.h"
+#include "vec4.h"
+#include "quat.h"
+#include "mat3.h"
+#include "mat4.h"
