--- 3D_Game_Engine/3D_Game_Engine/Math/JML.h
@@ -1,7 +1,0 @@
-
-#include "Math\Jade_Primitives.h"
-#include "Math\JadeMath.h"
-#include "Math\JadeCalculator.h"
-#include "Math\vec2.h"
-#include "Math\vec3.h"
-#include "Math\vec4.h"
