--- 3D_Game_Engine/3D_Game_Engine/Math/JadeCalculator.cpp
@@ -1,21 +1,0 @@
-#include "JadeCalculator.h"
-
-namespace jml
-{
-	//vec3 operator *(mat3 &param)
-	//{
-	//	return vec3(
-	//		(x * param.I11) + (y * param.I21) + (z * param.I31),  // x value
-	//		(x * param.I12) + (y * param.I22) + (z * param.I32),  // y value
-	//		(x * param.I13) + (y * param.I23) + (z * param.I33)); // z value
-	//}
-
-	//vec4 operator * (mat4 &param)
-	//{
-	//	return vec4(
-	//		(x * param.I11) + (y * param.I21) + (z * param.I31) + (w * param.I41),
-	//		(x * param.I12) + (y * param.I22) + (z * param.I32) + (w * param.I42),
-	//		(x * param.I13) + (y * param.I23) + (z * param.I33) + (w * param.I43),
-	//		(x * param.I14) + (y * param.I24) + (z * param.I34) + (w * param.I44);
-	//}
-}
