--- 3D_Game_Engine/3D_Game_Engine/Math/Jade_Primitives.h
@@ -1,28 +1,0 @@
-// ***********************************************************************
-// Author           : Jesse Derochie
-// Created          : 03-23-2016
-// Created For      : Jade Math Library
-// Last Modified By : Jesse Derochie
-// Last Modified On : 03-23-2016
-// ***********************************************************************
-// <copyright file="Jade_Primitives.h">
-//     Copyright (c) Jesse Derochie. All rights reserved.
-// </copyright>
-// <summary>JML Primitives header file</summary>
-// ***********************************************************************
-#pragma once
-
-namespace jml
-{
-	typedef signed char						JChar;
-	typedef unsigned char					U_JChar;
-	typedef signed short					JShort;
-	typedef unsigned short					U_JShort;
-	typedef signed int						JInt;
-	typedef unsigned int					U_JInt;
-	typedef signed __int64					JLongLong;
-	typedef unsigned __int64				U_JLongLong;
-	typedef float							JFloat;
-	typedef double							JDouble;
-	typedef bool							JBool;
-}
